+++ src/Manifest/index.ts
@@ -1,1 +1,2 @@
 export * from './Manifest.types'
+export * from './Manifest.router'
