+++ src/server.ts
@@ -2,6 +2,7 @@ import { env } from 'decentraland-commons'
 
 import { AppRouter } from './App'
 import { ProjectRouter } from './Project'
+import { ManifestRouter } from './Manifest'
 import { DeploymentRouter } from './Deployment'
 import { db } from './database'
 import { ExpressApp } from './common/ExpressApp'
@@ -21,6 +22,7 @@ app
 // Mount routers
 new AppRouter(app).mount()
 new ProjectRouter(app).mount()
+new ManifestRouter(app).mount()
 new DeploymentRouter(app).mount()
 
 /* Start the server only if run directly */
